requirements.txt
robomaster_gym
